renamed golang to goprint
