+++ print/goprint/script.go
@@ -1,4 +1,4 @@
-package golang
+package goprint
 
 import (
 	"github.com/gobuffalo/lush/ast"
